Use sam_itr_queryi() instead of formatting a region string
As we already have tid and nttable.beg/end are 0-based half open, use
sam_itr_queryi() instead of manufacturing a formatted region string.
Display the verbose message in 1-based human format -- note that the
previous `region` string was off-by-one (it was made as 0-based but
used as 1-based), both as logged and as used in sam_itr_querys().
Happily bam2R_pileup_function() restricts to [nttable.beg,nttable.end)
correctly so this made no difference to the results.

Also remove a couple of unused variables.