Filter by NM only on reads that do have an NM tag
BAM/CRAM files may have a mixture of reads with and without NM tags;
in particular, unmapped reads likely do not have NM tags. So checking
the first read does not give a definitive indication.

Instead introduce a function getNM() that returns the NM value if it
is present, or increments no_NM_count and returns a passes-filter value
if no NM tag is present. Use this function to write the two filter
checking expressions so that they do not crash on reads without NM.

Use no_NM_count to print a warning about lack of NM fields afterwards,
as appropriate. Remove the existing "get first read to check NM" code.
Re-opening the BAM/CRAM file was inelegant, and did not work with the
non-iterator branch as the re-opened file was not left in the correct
state for sam_read1().