Add ##date DataFrame exactly as Rsamtools would produce it
Ensure that the DataFrameList we add for the `##date` header is labelled
exactly as Rsamtools does for simple/unstructured header lines.

Since Bioconductor/VariantAnnotation#60 (released in 1.42.1), writeVcf()
uses the column name ("Value" for simple headers) to distinguish between
simple and structured header lines. Label the DataFrame accordingly so
that our `##date` header is recognised as a simple/unstructured header.